requirements.txt
apache-airflow==2.6.3
kaggle==1.5.15
pandas